+++ src/api/api.js
@@ -1,29 +1,4 @@
-import axios from 'axios';
-
-module.exports = (req, res) => {
-  const { name = 'World' } = req.query
-  res.status(200).send(`Hello ${name}!`)
-}
-
-const api = axios.create({
-	baseURL: '/api/',
+export default (request, response) =>
+  response.send({
+    data: 'hello world',
 });
-
-export const booksApi = {
-	search: item =>
-		api.get('search.api', {
-			params: {
-				key: process.env.REACT_APP_API_KEY,
-				query: item,
-				output: 'json',
-			},
-		}),
-	getList: category =>
-		api.get(`${category}.api`, {
-			params: {
-				key: process.env.REACT_APP_API_KEY,
-				categoryId: '100',
-				output: 'json',
-			},
-		}),
-};
