+++ src/routes/itemList/ItemListContainer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ItemListPresenter from './ItemListPresenter';
-import { booksApi } from 'api/api';
+import { booksApi } from 'api';
 
 class ItemListContainer extends React.Component {
 	state = {
